feat: enabling spookfox apps from elisp side
So users of spookfox can decide which spookfox functionality they want. I think
we should also allow configuring the apps from elisp itself. e.g I'd like to
access my firefox tabs without needing to write them to an org file.
