+++ spookfox-addon/src/manifest.json
@@ -30,5 +30,5 @@
     "icons/unchained-dark.svg"
   ],
 
-  "permissions": ["nativeMessaging", "tabs", "search"]
+  "permissions": ["tabs", "search", "scripting"]
 }
